ver2/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
matplotlib>=3.4.3
tqdm>=4.62.3
scikit-learn>=0.24.2
pillow>=8.3.1
tensorboard>=2.6.0 